pyproject.toml
[tool.poetry]
name = "movie-collector"
version = "0.1.0"
description = ""
authors = ["João Nogueira <joao.nogueira@datarisk.io>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.4.2"
requests = "^2.27.1"
tqdm = "^4.64.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
jupyter = "^1.0.0"
mypy = "^0.950"
pytest-cov = "^3.0.0"
flake8 = "^4.0.1"
pylint = "^2.13.9"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
